\name{RL1b-package}
\alias{RL1b-package}
\alias{RL1b}
\docType{package}
\title{
\packageTitle{RL1b}
}
\description{
\packageDescription{RL1b}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RL1b}
\packageIndices{RL1b}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RL1b}

Maintainer: \packageMaintainer{RL1b}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
